Class-based fact definition base was moved to dedicated package
