+++ yargy/shortcuts/__init__.py
@@ -1,0 +1,5 @@
+from .interpretation import FactDefinition
+
+__all__ = (
+    "FactDefinition",
+)
